+++ src/com/projects/gui/WorldStructuresPanel.java
@@ -33,7 +33,7 @@ public class WorldStructuresPanel extends StructurePanel implements SubscribedVi
 
             for (int i = 0; i < numberOfRows; ++i)
             {
-                if ((structureTable.getRow(i)).getId().intValue() == structure.getId().intValue())
+                if ((structureTable.getRow(i)).getId() == structure.getId())
                 {
                     structureTable.setRow(i, structure);
                 }
